requirements.txt
# torch            <-- ใส่ # ไว้
# torchvision      <-- ใส่ # ไว้
dvc[s3]
pandas
numpy
scikit-learn
opencv-python
timm
python-dotenv
pyyaml
boto3
transparent_background
Pillow